+++ packages/linejs/client/clients/internal/channel-client.ts
@@ -1,6 +1,6 @@
 // For Channel (channel, etc)
 
-import type { ProtocolKey } from "@evex/linejs-types/declares";
+import type { ProtocolKey } from "../../libs/thrift/declares.ts";
 import type * as LINETypes from "@evex/linejs-types";
 import { SquareClient } from "./square-client.ts";
