+++ packages/linejs/client/clients/internal/square-client.ts
@@ -1,8 +1,8 @@
 // For Square (square, etc)
 
-import type { NestedArray, ProtocolKey } from "@evex/linejs-types/declares";
+import type { NestedArray, ProtocolKey } from "../../libs/thrift/declares.ts";
 import type * as LINETypes from "@evex/linejs-types";
-import type { LooseType } from "@evex/linejs-types/entities";
+import type { LooseType } from "../../entities/common.ts";
 import { LiffClient } from "./liff-client.ts";
 
 export class SquareClient extends LiffClient {
