+++ packages/linejs/client/entities/common.ts
@@ -3,8 +3,14 @@
  * @module
  */
 
+/**
+ * Loose Type
+ */
 // deno-lint-ignore no-explicit-any
 export type LooseType = any;
 
+/**
+ * Always Type
+ */
 // deno-lint-ignore ban-types
 export type AlwaysType = {};
