+++ packages/types/deno.json
@@ -4,11 +4,6 @@
 	"license": "MIT",
 	"exports": {
 		".": "./line_types.ts",
-		"./entities": "./entities/index.ts",
-		"./read": "./read.ts",
-		"./declares": "./declares.ts",
-		"./thrift": "./thrift.ts",
-		"./parser": "./parser.ts",
-		"./write": "./write.ts"
+		"./thrift": "./thrift.ts"
 	}
 }
