--- packages/types/entities/index.ts
@@ -1,6 +1,0 @@
-/**
- * Export from the dir
- * @module
- */
-
-export * from "./common.ts";
